renamed all feed words to Home
