+++ lib/main.dart
@@ -12,7 +12,7 @@ import 'package:gameonconnect/view/theme/theme_provider.dart';
 import 'package:gameonconnect/view/pages/authentication/sign_up_page.dart';
 import 'view/pages/settings/help_page.dart';
 import 'package:provider/provider.dart';
-import 'view/pages/feed/feed_page.dart';
+import 'view/pages/home/home_page.dart';
 import 'view/pages/events/create_events_page.dart';
 import 'view/pages/events/view_events_page.dart';
 import 'firebase_options.dart';
@@ -99,7 +99,7 @@ class MyApp extends StatelessWidget {
               stream: FirebaseAuth.instance.authStateChanges(),
               builder: (BuildContext context, AsyncSnapshot<User?> snapshot) {
                 if (snapshot.hasData) {
-                  return FeedPage(
+                  return HomePage(
                     title: 'GameOnConnect',
                   );
                 } else {
@@ -107,7 +107,7 @@ class MyApp extends StatelessWidget {
                 }
               },
             ),
-        '/home': (context) => FeedPage(
+        '/home': (context) => HomePage(
               title: 'GameOnConnect',
             ),
         '/edit-profile': (context) => EditProfilePage(),
