+++ lib/view/components/home/connection_update.dart
