+++ lib/view/components/home/event_invite_list.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:gameonconnect/services/events_S/event_service.dart';
 import 'package:gameonconnect/model/events_M/events_model.dart';
-import 'package:gameonconnect/view/components/feed/event_invite.dart';
+import 'package:gameonconnect/view/components/home/event_invite.dart';
 
 class EventInvitesList extends StatefulWidget {
   const EventInvitesList({super.key});
