+++ lib/view/components/home/start_timer.dart
